IOS-8047 Fix review comments
