+++ BlockchainSdk/Extensions/String+.swift
@@ -117,12 +117,3 @@ extension String: @retroactive LocalizedError {
         return self
     }
 }
-
-extension String {
-    private static var numericSet: Set<Character> = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"]
-
-    var isNumeric: Bool {
-        guard !isEmpty else { return false }
-        return Set(self).isSubset(of: String.numericSet)
-    }
-}
